+++ lib/graphql/iterableModel.ts
@@ -1,19 +1,25 @@
 import { ResultPage } from "./resultPage";
 import { Model } from "./model";
+import { FetchOptions } from "./types";
 
-export abstract class IterableModel<T> extends Model<T> {
-  [Symbol.asyncIterator]() {
-    let lastResult!: ResultPage<T>;
+export abstract class IterableModel<
+  ModelType,
+  FetchOptionsType = FetchOptions
+> extends Model<ModelType, FetchOptionsType> {
+  createAsyncIterator(args: FetchOptionsType) {
+    const fetch = this.fetch.bind(this);
+    let lastResult!: ResultPage<ModelType, FetchOptionsType>;
 
     return {
       next: async () => {
         // this is the first call or items are empty
         if (!lastResult) {
-          lastResult = await this.fetch();
+          lastResult = await fetch(args);
         }
 
         if (lastResult.items.length === 0 && lastResult.pageInfo.hasNextPage) {
-          lastResult = await this.fetch({
+          lastResult = await fetch({
+            ...args,
             after: lastResult.pageInfo.endCursor
           });
         }
@@ -28,6 +34,15 @@ export abstract class IterableModel<T> extends Model<T> {
 
         // no more data
         return { value: undefined, done: true };
+      }
+    };
+  }
+
+  fetchAll(args: FetchOptionsType) {
+    const asyncIterator = this.createAsyncIterator(args);
+    return {
+      [Symbol.asyncIterator]() {
+        return asyncIterator;
       }
     };
   }