+++ tests/graphql/transaction.spec.ts
@@ -2,19 +2,27 @@ import { expect } from "chai";
 import * as sinon from "sinon";
 import { RawQueryResponse, SubscriptionType } from "../../lib/graphql/types";
 import { Transaction } from "../../lib/graphql/schema";
+import { Client } from "../../lib";
 import { createClient, createTransaction } from "../helpers";
 import { NEW_TRANSACTION_SUBSCRIPTION } from "../../lib/graphql/transaction";
 
 describe("Transaction", () => {
   describe("iterator", () => {
-    it("can iterate on all user transactions", async () => {
-      const client = createClient();
-      const stub = sinon.stub(client.graphQL, "rawQuery");
+    let client: Client;
+    let firstTransaction: any;
+    let secondTransaction: any;
+    let firstResponse: any;
+    let secondResponse: any;
+    let stub: any;
 
-      const firstTransaction = createTransaction();
-      const secondTransaction = createTransaction();
+    beforeEach(() => {
+      client = createClient();
+      stub = sinon.stub(client.graphQL, "rawQuery");
 
-      const firstResponse: any = {
+      firstTransaction = createTransaction();
+      secondTransaction = createTransaction();
+
+      firstResponse = {
         viewer: {
           mainAccount: {
             transactions: {
@@ -37,7 +45,7 @@ describe("Transaction", () => {
 
       stub.onFirstCall().resolves(firstResponse);
 
-      const secondResponse: any = {
+      secondResponse = {
         viewer: {
           mainAccount: {
             transactions: {
@@ -59,13 +67,52 @@ describe("Transaction", () => {
       };
 
       stub.onSecondCall().resolves(secondResponse as RawQueryResponse);
+    });
 
+    afterEach(() => {
+      stub.restore();
+    });
+
+    it("can fetch next page using the nextPage method", async () => {
+      const firstPage = await client.models.transaction.fetch({
+        first: 1
+      });
+
+      expect(typeof firstPage.nextPage).to.equal("function");
+      expect(firstPage.items).to.deep.equal([firstTransaction]);
+
+      const secondPage = firstPage.nextPage && (await firstPage.nextPage());
+      expect(secondPage?.items).to.deep.equal([secondTransaction]);
+    });
+
+    it("can iterate on all user transactions using the fetchAll iterator", async () => {
       let transactions: Array<Transaction> = [];
-      for await (const transaction of client.models.transaction) {
+      for await (const transaction of client.models.transaction.fetchAll()) {
         transactions = transactions.concat(transaction as Transaction);
       }
 
       expect(transactions).to.deep.equal([firstTransaction, secondTransaction]);
+    });
+
+    describe("when iterating backwards", () => {
+      it("can fetch the previous page using the previousPage method", async () => {
+        firstResponse.viewer.mainAccount.transactions.pageInfo.hasNextPage = false;
+        secondResponse.viewer.mainAccount.transactions.pageInfo.hasPreviousPage = true;
+
+        stub.onFirstCall().resolves(secondResponse);
+        stub.onSecondCall().resolves(firstResponse);
+
+        const firstPage = await client.models.transaction.fetch({
+          last: 1
+        });
+
+        expect(typeof firstPage.previousPage).to.equal("function");
+        expect(firstPage.items).to.deep.equal([secondTransaction]);
+
+        const secondPage =
+          firstPage.previousPage && (await firstPage.previousPage());
+        expect(secondPage?.items).to.deep.equal([firstTransaction]);
+      });
     });
   });