+++ tests/helpers.ts
@@ -1,4 +1,9 @@
-import { TransactionProjectionType, Transaction } from "../lib/graphql/schema";
+import {
+  TransactionProjectionType,
+  Transaction,
+  Transfer,
+  TransferStatus
+} from "../lib/graphql/schema";
 import { Client } from "../lib";
 
 export const clientId = "26990216-e340-4f54-b5a5-df9baacc0440";
@@ -25,5 +30,24 @@ export const createTransaction = (): Transaction => {
     fees: [],
     bookingDate: new Date(),
     paymentMethod: "card"
+  };
+};
+
+export const createTransfer = (
+  override: Record<string, any> = {}
+): Transfer => {
+  return {
+    id: Math.random.toString(),
+    recipient: "John Doe",
+    iban: "DE32110101001000000029",
+    amount: parseInt((Math.random() * 100).toString()),
+    status: TransferStatus.Confirmed,
+    executeAt: null,
+    lastExecutionDate: null,
+    purpose: "some transfer purpose",
+    e2eId: "some-e2e-id",
+    reoccurrence: null,
+    nextOccurrence: null,
+    ...override
   };
 };