+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Introduce a method to subscribe to new transactions (`models.transaction.subscribe`)
+- Introduce methods for fetching transfers (`models.transfer.fetch`, `models.transfer.fetchAll`)
+
+### Changed
+- *Breaking*: changed transaction async iterator from `models.transaction` to `models.transaction.fetchAll`.
 
 ## [0.21.0] - 2019-11-22
 ### Changed